+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/dto/response/GoogleUserInfoResponse.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.account.auth.oauth.dto.response;
 
 import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
-import com.mobble.mobbleserver.global.exception.errorCode.oAuth2.OAuth2ErrorCode;
+import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
 
 import java.util.Map;
 
@@ -13,7 +13,7 @@ public class GoogleUserInfoResponse implements OAuth2UserInfo {
     private final SocialProvider socialProvider = SocialProvider.GOOGLE;
 
     public GoogleUserInfoResponse(Map<String, Object> attributes) {
-        if (attributes == null || attributes.get("sub") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (attributes == null || attributes.get("sub") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
         this.attributes = attributes;
     }
 
@@ -29,14 +29,14 @@ public class GoogleUserInfoResponse implements OAuth2UserInfo {
 
     @Override
     public String getName() {
-        if (attributes.get("name") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (attributes.get("name") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
 
         return attributes.get("name").toString();
     }
 
     @Override
     public String getEmail() {
-        if (attributes.get("email") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (attributes.get("email") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
 
         return attributes.get("email").toString();
     }
