+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/dto/response/KakaoUserInfoResponse.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.account.auth.oauth.dto.response;
 
 import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
-import com.mobble.mobbleserver.global.exception.errorCode.oAuth2.OAuth2ErrorCode;
+import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
 
 import java.util.Map;
 
@@ -13,7 +13,7 @@ public class KakaoUserInfoResponse implements OAuth2UserInfo {
     private final SocialProvider socialProvider = SocialProvider.KAKAO;
 
     public KakaoUserInfoResponse(Map<String, Object> attributes) {
-        if (attributes == null || attributes.get("id") == null || attributes.get("kakao_account") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (attributes == null || attributes.get("id") == null || attributes.get("kakao_account") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
         this.attributes = attributes;
     }
 
@@ -30,7 +30,7 @@ public class KakaoUserInfoResponse implements OAuth2UserInfo {
     @Override
     public String getName() {
         Map<String, Object> kakaoAccount = (Map<String, Object>) attributes.get("kakao_account");
-        if (kakaoAccount == null || kakaoAccount.get("name") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (kakaoAccount == null || kakaoAccount.get("name") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
 
         return kakaoAccount.get("name").toString();
     }
@@ -38,7 +38,7 @@ public class KakaoUserInfoResponse implements OAuth2UserInfo {
     @Override
     public String getEmail() {
         Map<String, Object> kakaoAccount = (Map<String, Object>) attributes.get("kakao_account");
-        if (kakaoAccount == null || kakaoAccount.get("email") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (kakaoAccount == null || kakaoAccount.get("email") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
 
         return kakaoAccount.get("email").toString();
     }
