+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/dto/response/NaverUserInfoResponse.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.account.auth.oauth.dto.response;
 
 import com.mobble.mobbleserver.account.auth.oauth.service.SocialProvider;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
-import com.mobble.mobbleserver.global.exception.errorCode.oAuth2.OAuth2ErrorCode;
+import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
 
 import java.util.Map;
 
@@ -13,7 +13,7 @@ public class NaverUserInfoResponse implements OAuth2UserInfo {
     private final SocialProvider socialProvider = SocialProvider.NAVER;
 
     public NaverUserInfoResponse(Map<String, Object> rawResponse) {
-        if (rawResponse == null || rawResponse.get("response") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (rawResponse == null || rawResponse.get("response") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
         this.response = (Map<String, Object>) rawResponse.get("response");
     }
 
@@ -24,21 +24,21 @@ public class NaverUserInfoResponse implements OAuth2UserInfo {
 
     @Override
     public String getProviderId() {
-        if (response.get("id") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (response.get("id") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
 
         return response.get("id").toString();
     }
 
     @Override
     public String getName() {
-        if (response.get("name") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (response.get("name") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
 
         return response.get("name").toString();
     }
 
     @Override
     public String getEmail() {
-        if (response.get("email") == null) throw new DomainException(OAuth2ErrorCode.NO_USER_INFO);
+        if (response.get("email") == null) throw new DomainException(OAuthErrorCode.NO_USER_INFO);
 
         return response.get("email").toString();
     }
