+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/verifier/AbstractSocialTokenVerifier.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.account.auth.oauth.verifier;
 
 import com.mobble.mobbleserver.account.auth.oauth.verifier.dto.SocialUserInfo;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
-import com.mobble.mobbleserver.global.exception.errorCode.oAuth2.OAuth2ErrorCode;
+import com.mobble.mobbleserver.global.exception.errorCode.oAuth.OAuthErrorCode;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatusCode;
@@ -28,11 +28,11 @@ public abstract class AbstractSocialTokenVerifier implements SocialVerifier {
                     .header(HttpHeaders.AUTHORIZATION, "Bearer " + accessToken)
                     .retrieve()
                     .onStatus(HttpStatusCode::isError, (req, res) -> {
-                        throw new DomainException(OAuth2ErrorCode.INVALID_ACCESS_TOKEN);
+                        throw new DomainException(OAuthErrorCode.INVALID_ACCESS_TOKEN);
                     })
                     .body(Map.class);
         } catch (Exception e) {
-            throw new DomainException(OAuth2ErrorCode.FAILED_TO_REQUEST_USER_INFO);
+            throw new DomainException(OAuthErrorCode.FAILED_TO_REQUEST_USER_INFO);
         }
         
         return parseUserInfo(response);
